requirements.txt
# Forex Radar Bot & AI Screener - Dependencies
# Install dengan: pip install -r requirements.txt

# === Data Market ===
yfinance>=0.2.40

# === Analisa Teknikal ===
pandas>=2.0.0
ta>=0.11.0
numpy>=1.24.0

# === Chart Generation ===
mplfinance>=0.12.10b0
matplotlib>=3.8.0

# === Web Scraping / Berita ===
feedparser>=6.0.10
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# === AI / Gemini API (fallback) ===
google-genai>=0.8.0

# === AI / Groq API (primary - faster, higher rate limit) ===
groq>=0.11.0

# === Telegram Bot ===
python-telegram-bot[job-queue]>=20.7

# === Utilitas ===
python-dotenv>=1.0.0
pytz>=2024.1
aiohttp>=3.9.0
aiosqlite>=0.20.0
